+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Release Notes
 
 ## Upcoming
+- Dropped support for PHP 8.0 
 - Removed support for prop-types
 - Removed support for class-based components
 - Removed semicolons from stubs
+- Upgraded to PHPUnit 10
 
 ## v3.0.0
 - Support Laravel 10